REFACTOR: 뷰 관련 로직 ViewModel로 분리
